+++ NEWS.md
@@ -11,7 +11,7 @@
 * checkmates to use `sptf_parameters` instead of hardcoding min and maxvalues in
 each function
 * export whc and pmn ptf's in NAMESPACE
-* `sptf_whc4` now takes A_DEPTH in m instead of cm 
+* `sptf_whc4` and `sptf_whc7` now takes A_DEPTH in m instead of cm 
 
 
 ## Version 0.1.0  2022-08-15